requirements.txt
streamlit
langchain
pymongo
fastapi
uvicorn
python-dotenv
requests
PyGithub